+++ tap_facebook/exceptions.py
@@ -1,0 +1,7 @@
+from facebook_business.exceptions import FacebookRequestError
+
+class TapFacebookException(Exception):
+    pass
+
+class InsightsJobTimeout(TapFacebookException):
+    pass
